requirements.txt
requests == 2.31.0
streamlit == 1.41.0
pillow == 11.0.0